pyproject.toml
[tool.poetry]
name = "app-event-management"
version = "0.1.0"
description = "A sample event management application"
authors = ["Shi Han NG <shihanng@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.5"
djangorestframework = "^3.12.2"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.8.4"
pylint = "^2.6.0"
mypy = "^0.790"
pytest = "^6.2.1"
pytest-django = "^4.1.0"
coverage = "^5.3.1"
flake8-isort = "^4.0.0"
factory-boy = "^3.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
